8193.37-13 Support (#1804)

+++ NWNXLib/API/API/CEffectIconObject.hpp
@@ -14,7 +14,7 @@ typedef int BOOL;
 
 struct CEffectIconObject
 {
-    uint8_t m_nIcon;
+    uint16_t m_nIcon;
     BOOL m_bFlashing;
     BOOL m_nPlayerBar;
